+++ src/main/java/ca/corefacility/bioinformatics/irida/validators/annotations/validators/LongitudeValidator.java
@@ -17,8 +17,8 @@ public class LongitudeValidator implements ConstraintValidator<Longitude, String
 
 	private static final Pattern LONGITUDE_PATTERN = Pattern.compile("^-?(\\d){3}(\\.\\d+)?$");
 
-	private static final Float LONG_MIN = -180f;
-	private static final Float LONG_MAX = 180f;
+	private static final Double LONG_MIN = -180d;
+	private static final Double LONG_MAX = 180d;
 
 	@Override
 	public void initialize(Longitude constraintAnnotation) {
@@ -33,7 +33,7 @@ public class LongitudeValidator implements ConstraintValidator<Longitude, String
 		}
 
 		if (LONGITUDE_PATTERN.matcher(value).matches()) {
-			Float longitude = Float.valueOf(value);
+			Double longitude = Double.valueOf(value);
 			return longitude >= LONG_MIN && longitude <= LONG_MAX;
 		}
